extract useWindowWidth custom hook to its own file
